+++ npc/cli/core.py
@@ -165,6 +165,14 @@ def _make_parser():
     parser_changeling.add_argument('-m', '--motley', help="The character's Motley", metavar='motley')
     parser_changeling.set_defaults(func=commands.create_changeling)
 
+    # Subcommand for making changelings, with their unique options
+    parser_werewolf = subparsers.add_parser('werewolf', aliases=['c'], parents=[common_options, character_options], help="Create a new werewolf character")
+    parser_changeling.add_argument('name', help="Character name", metavar='name')
+    parser_changeling.add_argument('auspice', help="The character's Auspice", metavar='auspice')
+    parser_changeling.add_argument('-t', '--tribe', help="The character's Tribe", metavar='tribe')
+    parser_changeling.add_argument('-p', '--pack', help="The character's Pack", metavar='pack')
+    parser_changeling.set_defaults(func=commands.create_changeling)
+
     # Subcommand for linting characer files
     parser_lint = subparsers.add_parser('lint', parents=[common_options, paths_parser], help="Check the character files for minimum completeness")
     parser_lint.add_argument('-f', '--fix', action='store_true', default=False, help="Automatically fix certain problems")